1. Making sure the model-container compiles on my laptop. 2. Disabled the compilation of the python bindings.

+++ neurospaces/idin.h
@@ -369,7 +369,7 @@ int IdinIsPrintable(struct symtab_IdentifierIndex *pidin)
 /*     if ((int)pidin & 1) */
 /* #endif */
 
-    if ((int)pidin & 1) 
+    if ((long int)pidin & 1) 
     {
 	return(FALSE);
     }
